+++ src/sensors/i2c_sensors/measurables/temperature_measurable.h
@@ -4,10 +4,10 @@
 #include <src/sensors/i2c_sensors/measurables/i2c_measurable.h>
 #include <src/sensors/i2c_sensors/mcp9808.hpp>
 
-class TemperatureMeasurable : public I2cMeasurable<double>{
+class TemperatureMeasurable : public I2cMeasurable<float> {
    public:
     TemperatureMeasurable(MCP9808* temp_sensor);
-    double TakeI2cReading();
+    float TakeI2cReading();
 };
 
 #endif  // SRC_SENSORS_I2C_SENSORS_MEASURABLES_TEMPERATURE_MEASURABLE_H_
